requirements.txt
# Core Dash dependencies
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tables>=3.8.0  # For HDF5 support

# API and data sources
requests>=2.31.0
python-dotenv>=1.0.0
toml>=0.10.2

# Time series analysis dependencies
statsmodels>=0.14.0  # For advanced time series analysis
scipy>=1.10.0  # For statistical functions
scikit-learn>=1.3.0  # For preprocessing and ML features

# Optional but recommended
openpyxl>=3.1.0  # For Excel file support

# External API packages (if available)
# myeia  # Uncomment if you have access
# nasspython  # Uncomment if you have access